Change container and action_bar_container when selecting navbar subreddit
PlaceholderFragment was renamed to StatusBarFragment. Also added some lines in
layouts so that each attribute is on its own line (for readability).
NavigationDrawerFragment listview selection switched from position:int to
subreddit:String as this is more useful to pass around between activities. The
front page needed to be switched to "/" to avoid conflict with the subreddit
/r/frontpage. The example action is hidden from the action bar